requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
discord.py==2.4.0
frozenlist==1.4.1
idna==3.10
immutabledict==4.2.0
minify_html==0.15.0
multidict==6.1.0
numpy==2.1.2
ortools==9.11.4210
pandas==2.2.3
propcache==0.2.0
protobuf==5.26.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
tzdata==2024.2
yarl==1.16.0
